Simplified Cookie Deletion Process
To eliminate future duplicate code a function called
deleteSemesterCookie was created in views.py to serve as the semester
cookie deleting function for any and all pages that might need this
option. An explanation of how to form links to utilize this function is
included in a comment in the function heading.